requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-genai==0.16.0  # ou a versão que você usou
PyPDF2==3.0.1
python-docx==1.1.2
werkzeug==3.0.3